**Refactor Journal Entry Handling:** * Created helper files for:     * `cleanupUploads`: Manages temporary file cleanup.     * `journalEntryValidation`: Centralizes journal entry validation.     * `cleanupCloudinaryData`: Handles Cloudinary resource deletion. * Refactored `addOrUpdateJournal` controller to utilize the new helper structure. * Implemented delete journal route and controller. * Modified `cleanupCloudinaryData` to accept the journal object instead of `userId` for more efficient cleanup.
