+++ backend/src/routes/journalRoutes.js
@@ -4,6 +4,7 @@ const {
   addOrUpdateJournal,
   getAllJournal,
   journalByDate,
+  deleteJournal,
 } = require("../controllers/journalController");
 const { protect } = require("../middleware/authMiddleware");
 const { upload } = require("../middleware/multerCloudinaryMiddleware");
@@ -12,5 +13,6 @@ const { upload } = require("../middleware/multerCloudinaryMiddleware");
 router.post("/addOrUpdate", protect, upload, addOrUpdateJournal);
 router.get("/:date", protect, upload, journalByDate);
 router.get("/getJournals", protect, getAllJournal);
+router.post("/deleteJournal", protect, deleteJournal);
 
 module.exports = router;
